Applied Better Practices for file-naming
