+++ lib/cnn_news_bot/commands/get_news.rb
@@ -1,7 +1,6 @@
-# rubocop:disable Security/Open
-
 require 'rss'
 require 'open-uri'
+require_relative '../classes/slack_interaction'
 
 module CnnNews
   module Commands
@@ -36,35 +35,10 @@ module CnnNews
 
       command 'science' do |client, data, _match|
         url = 'http://rss.cnn.com/rss/edition_space.rss'
-        
+
         science_obj = SlackInteraction.new(url)
         science_obj.interact(client, data.channel)
       end
     end
   end
 end
-
-class SlackInteraction
-  attr_reader :url
-
-  def initialize(url)
-    @url = url
-  end
-  
-  def interact(client, to_channel)
-    speaker(client, to_channel, url_parser)
-  end
-
-  private
-  def url_parser
-    RSS::Parser.parse(open(@url).read, false)
-  end
-
-  def speaker(client, to_channel, info)
-    client.say(channel: to_channel, text: info.items.first.link)
-    client.say(channel: to_channel, text: info.items.second.link)
-    client.say(channel: to_channel, text: info.items.third.link)
-  end
-end
-
-# rubocop:enable Security/Open
